requirements.txt
fastapi
uvicorn
sqlalchemy
jinja2
websockets
python-multipart
python-dotenv
tree_sitter==0.25.2
tree_sitter_python==0.25.0
tree_sitter_c==0.24.1 
tree_sitter_cpp==0.23.4 
tree_sitter_go==0.25.0 
tree_sitter_java==0.23.5 
tree_sitter_javascript==0.25.0 
tree_sitter_typescript==0.23.2
neo4j==5.28.2
python-dotenv
ollama
neo4j-graphrag